requirements.txt
python-telegram-bot==21.7
requests==2.31.0
pg8000==1.30.3
python-dotenv==1.0.0
